#3215 - Reorganize CAS storage configuration properties
- Renamed CasStorageProperties to CasStorageCacheProperties
- Introduce new CasStorageProperties where the compression for CASes can be configured
- Update documentation
