+++ inception/inception-api/src/main/java/de/tudarmstadt/ukp/clarin/webanno/api/config/RepositoryProperties.java
@@ -21,7 +21,6 @@ import java.io.File;
 
 import org.slf4j.MDC;
 import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.jmx.export.annotation.ManagedAttribute;
 import org.springframework.jmx.export.annotation.ManagedResource;
 
 import de.tudarmstadt.ukp.clarin.webanno.support.SettingsUtil;
@@ -37,44 +36,6 @@ import de.tudarmstadt.ukp.clarin.webanno.support.logging.Logging;
 public class RepositoryProperties
 {
     private File path;
-
-    private boolean compressedCasSerialization = false;
-    private boolean paranoidCasSerialization = false;
-    private boolean traceAccess = false;
-
-    public void setTraceAccess(boolean aTraceAccess)
-    {
-        traceAccess = aTraceAccess;
-    }
-
-    public boolean isTraceAccess()
-    {
-        return traceAccess;
-    }
-
-    @ManagedAttribute
-    public void setParanoidCasSerialization(boolean aParanoidCasSerialization)
-    {
-        paranoidCasSerialization = aParanoidCasSerialization;
-    }
-
-    @ManagedAttribute
-    public boolean isParanoidCasSerialization()
-    {
-        return paranoidCasSerialization;
-    }
-
-    @ManagedAttribute
-    public void setCompressedCasSerialization(boolean aCompressedCasSerialization)
-    {
-        compressedCasSerialization = aCompressedCasSerialization;
-    }
-
-    @ManagedAttribute
-    public boolean isCompressedCasSerialization()
-    {
-        return compressedCasSerialization;
-    }
 
     public File getPath()
     {
