--- inception/inception-doc/src/main/resources/META-INF/asciidoc/admin-guide/settings_repository.adoc
@@ -1,33 +1,0 @@
-// Licensed to the Technische Universität Darmstadt under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The Technische Universität Darmstadt 
-// licenses this file to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.
-//  
-// http://www.apache.org/licenses/LICENSE-2.0
-// 
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-= Repository Settings
-
-These settings relate to the storage of data on the file system - mostly annotation data.
-
-.Server settings
-[cols="4*", options="header"]
-|===
-| Setting
-| Description
-| Default
-| Example
-
-| repository.compressed-cas-serialization
-| Whether to compress annotation files
-| false
-| true
-|===
