+++ inception/inception-export/src/test/java/de/tudarmstadt/ukp/inception/export/DocumentImportExportServiceImplTest.java
@@ -73,6 +73,7 @@ import de.tudarmstadt.ukp.clarin.webanno.api.dao.casstorage.CasStorageServiceImp
 import de.tudarmstadt.ukp.clarin.webanno.api.dao.casstorage.CasStorageSession;
 import de.tudarmstadt.ukp.clarin.webanno.api.dao.casstorage.FileSystemCasStorageDriver;
 import de.tudarmstadt.ukp.clarin.webanno.api.dao.casstorage.config.BackupProperties;
+import de.tudarmstadt.ukp.clarin.webanno.api.dao.casstorage.config.CasStorageCachePropertiesImpl;
 import de.tudarmstadt.ukp.clarin.webanno.api.dao.casstorage.config.CasStoragePropertiesImpl;
 import de.tudarmstadt.ukp.clarin.webanno.api.type.CASMetadata;
 import de.tudarmstadt.ukp.clarin.webanno.model.AnnotationFeature;
@@ -114,10 +115,10 @@ public class DocumentImportExportServiceImplTest
         MDC.put(Logging.KEY_REPOSITORY_PATH, repositoryProperties.getPath().toString());
 
         CasStorageDriver driver = new FileSystemCasStorageDriver(repositoryProperties,
-                new BackupProperties());
+                new BackupProperties(), new CasStoragePropertiesImpl());
 
         CasStorageServiceImpl storageService = new CasStorageServiceImpl(driver, null, null,
-                new CasStoragePropertiesImpl());
+                new CasStorageCachePropertiesImpl());
 
         sut = new DocumentImportExportServiceImpl(repositoryProperties,
                 List.of(new XmiFormatSupport()), storageService, schemaService, properties);
